Explicitly reference directory index (fixes #1)
Node (and Typescript in node module resolution mode) when asked to load a directory name will try loading an index file in that directory (index.js for Node; index.ts and then index.d.ts for Typescript in node resolution mode). Typings compiler doesn't current support this sort of resolution (see typings/typings#173). This is the easiest fix for the issue.